requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
anthropic>=0.28.0
pdfplumber>=0.11.0
